fix event mixing seg fault, remove Mixer child class
The SimplerMixerOperator has been removed and the class constructor
paramerers of HemisphereMixerOperator have been simplified. Probably
segfault was somewhere in the lambda that was used in the
initialization list. The only drawback right now is that the distance
function can only be subset during runtime and new functions have to
be added at compile time.